Transport Movement type is added to searchFind shape file.
In event a mistake related to the class EventDateTime is corrected. Domain and Range were created for the class and now it is now corrected. 

The DigitalTwin re-loaded correctly. 

In Classifications.ttl the prefix for DigitalTwin is corrected,
The prefix for businessService is corrected.  

In BusinessService.ttl the DigitalTwin prefix is corrected for a syntax error.